stream-processor/requirements.txt
pyspark==3.4.0
psycopg2-binary==2.9.7
kafka-python==2.0.2